Avoid blocking read actions
AnnotatedMethodsFinder#getContainingUMethodAnnotatedWith would still try
to get a blocking read lock even when it should be running under a non
blocking one. The method was already tagged as @RequiredReadLock so it
should not be getting the lock itself.
This removes the lock if not needed and logs an error if the method is
called without it (which should not happen).

Test: Existing
Bug: 349535556
Change-Id: Ie4a703a67b67214486de5bb06db7d61c375997f6

GitOrigin-RevId: 0c56da0df2c832df550cb7b3b99e1b3a6c5fab00