Add type parameters to AnimationState interface
Introduces type parameters to the AnimationState interface and update
the implementations accordingly. This allows for better type safety when
working with different animation states in the Compose Animation
Inspector.

Replaces the `stateHashCode` with a `state` property that directly holds
the current state of the animation.
It relies on the fact that .hashCode and equals will be correctly
implemented for state. Even if not, there is not going to be user visible
bugs, only extra loading of transition of state change.

Adds FromToSupportedAnimationManager, new class designed to handle
Compose animations that have distinct "from" and "to" states.

UpdateFromAndToStates functions in AnimationClock.kt are modified to
accept lists, because it's what we expect on androidx side

Fixes: 348574356
Test: *StateTest
Change-Id: If5a6771cba3d5975bb526b7a4051ec365f85a8e9

GitOrigin-RevId: b5b087ef34525da7c36e51592540244de41f10bd