Handle SequentialCollection in ApiDetector
Warn about upcoming rebinding of extension functions, and when there
is a new API warning (when using API level 35) add quickfix to rewrite
to safer alternative.

(The IDE-side fix of this is basically updating the API-relate
inspection, which adds various IDE-implemented quickfixes related
to API constraints, to not only handle a single LintFix map descriptor
which supplies the API constraint associated with a particular warning,
but to also allow a normal quickfix to be attached. We use that here
to offer a quickfix for these two new methods to replace the new
platform SequentialCollection methods to be replaced by safe Kotlin
extension replacements.)

Test: Unit test included
Bug: 320460092
Change-Id: I9fbd4478f81843171d0f3a925253c3260d1af0c2

GitOrigin-RevId: 2e82d48df39cb837679f7ed133ccd286236aec88