Navigate to the preview definition when single clicking
... on a CommonPreviewRepresentation preview.

This is to be more in line with the behaviour in Compose previews.
Ideally, we would want to navigate to the view component that is being
clicked on. However, that requires mapping the x/y coordinates to the
view component, which we currently do not have.

In the case of tile previews, that would also require mapping the
inflated view components to the original proto layout component.

In the meantime, navigating to the preview definition is better than
having no navigation at all.

Fixes: 347711875
Test: CommonPreviewRepresentationTest
Change-Id: I60d3571cfbb4da99e0a8ae6168182150de371aa5

GitOrigin-RevId: 90abc2fe41b005810def3744c31fdd4aaf6689e5