Add isDataBindingEnabled to AndroidModuleSystem
Implement it for Gradle by reading from the GradleAndroidModel.  For
other module systems, have a default get() returning false.

Use it, rather than inspecting dependencies, in the databinding mode
determination in LayoutBindingModuleCache.  (Adjust a test which
verifies that the determination is correct for Gradle to have the
right error message on failure).

Delete a test that tested nothing other than the previous
implementation's implementation strategy.  As a bonus, this also
deletes some (non-production) usages of GradleVersion.

Fixes: 350722344
Test: existing tests adjusted
Change-Id: I6a3d161c3990343e836cb3c527b3547e299abf3f

GitOrigin-RevId: 33bfcf7ea06e35be1d823019a05362c156347c43