Open projects waits for AndroidGradleProjectStartupActivity on tests
With Intellij 242 the ProjectActivity are no longer awaited on
project open during test execution, causing AndroidGradleTestUtils
to PlatformTestUtil.dispatchAllEventsInIdeEventQueue() before
AndroidGradleProjectStartupActivity was initialized. This results
on tests finishing execution before Gradle sync was triggered.

To avoid this, AndroidGradleTestUtils.openPreparedProject
will explicitly wait for AndroidGradleProjectStartupActivity to
be initialized before dispatching allEvents in the queue.

Bug: 351913178
Test: UnusedResourcesGradleTest
Change-Id: Idb1c4b2f87ccdc50ba2a1b9be5e44d1e1ecf4f25

GitOrigin-RevId: 53668757bcb19e50e09d5886a045dd96f09ba5d7