Convert ProjectSystemProviders to application extension points
The main difficulty here is to provide the needed idempotence for
determining the project system without introducing unnecessary
locking.

The code in ProjectSystemService is a hand-rolled implementation of
the Kotlin LazyThreadSafetyMode.PUBLICATION mode, so convert to using
that.  Preserving the ability for tests to override the project system
is, as a result, somewhat uglier, but at least it remains very
explicitly test-only.

This does not address the use in some underlying project system
providers of (implicitly) LazyThreadSafetyMode.SYNCHRONIZED to
guarantee the construction of only one instance.  Since we no longer
have a project-level object on which we can hang the project system,
instead add the requirement that project system constructors must be
idempotent.  Then, we can construct as many objects as we like, and
all but one will be discarded and garbage collected.

This requires us not to leak the project system instance, for example
by connecting it to a message bus in its creation.  To avoid this,
make sure that any state required (such as a message bus connection)
is in a project-level service.

Bug: 352000992
Test: existing
Change-Id: I9b7ceca85d1f8861ea66a699beb66c526d40f47c

GitOrigin-RevId: 80ee7d669064895a3389148ec8fa04d9fec73603