New MethodInterceptTransform
A class bytecode transform that allows to intercept any method calls.
The transforms will intercept static and instance invocations and will
call a trampoline method with the parameters that were being sent to the
call.
The trampoline method can just exit to let the flow continue or throw
an exception.

This transform is being introduced to build the render sandbox where a
large number of methods need to be intercepted. Since the sandbox will
require many diverse methods to intercept, it's not practical to create
a manual intercept system for each one.

Bug: N/A
Test: MethodInterceptTransformTest
Change-Id: I1e949895b99363e699403c87f2d368226231b481

GitOrigin-RevId: 214905c9c3c0533ab23253d9596ce9ace2f3cd97