Do not call PsiAwareTextEditorProvider#createEditorBuilder with read lock
When initializating, SourceCodeEditorProvider#createEditorAsync is
called while holding the readlock. The method will in turn invoke
createEditorBuilder from PsiAwareTextEditorProvider. This method can
deadlock if AWT is waiting for the write lock since the caller was
holding the read lock.

This change moves the SourceCodeEditorProvider builder initialization to
createEditorBuilder. This is still marked as experimental but is the
equivalent to createEditorAsync WITHOUT holding the read lock
(createEditorAsync just gets the read lock and calls
createEditorBuilder).

Test: SourceCodeEditorProviderTest
Fixes: 352345934
Change-Id: I707171c1ad683dce7c2fe5184f12a5d4df567e86

GitOrigin-RevId: b61118ad369e188a48872c4cc53b197f1de34532