Separate AA usage from write-action of ComposeVisualLintIssueProvider for K2
ComposeVisualLintIssueProvider uses `fqNameMatches(..)` on a
write-action that uses analysis APIs (AA). Since running the analysis
API on a write-action can cause wrong analysis results (imagine the PSI
is updated, which makes the analysis invalid), K2 restricts the AA use
on a write-action. This commit separates `fqNameMatches(..)` from the
write-action of ComposeVisualLintIssueProvider for K2.

Bug: 348627882
Test: ComposeVisualLintSuppressTaskTest.testSuppressIssue
Change-Id: Ibecf6bb1771aee24a74c66874f2b0f364aeda07b

GitOrigin-RevId: 4d1f431aa1b0c470f46a257a2d425110373d170a