Move DaggerRelatedItem...Test off EDT
Some of the actions involved can no longer be called on the EDT after
the 2024.2 merge because they involve Kotlin analysis. This is not a
problem in the real product, and only happens in tests.

This test also exposed an issue with caching that only appears to happen
in K2. I'm disabling the DAGGER_CACHE_RELATED_ELEMENTS flag since it's
not essential, and b/352564637 tracks investigating and re-enabling that
flag.

Test: DaggerRelatedItemLineMarkerProviderTest
Fixes: 352544230
Bug: 352564637
Change-Id: I450350f486559d0ef628c54bcea10fa84a9c866c

GitOrigin-RevId: c3dc915a651cebfbb8067408e919ce300a5cfc93