Use compiler error context experiment flag in ExplainSyncOrBuildOutput
This change adds the StudioBotExternalFlags extension point in ml-api,
which should be used to expose flags that are controlled by aiplugin but
need to be used by features in e.g. tools/adt/idea. In this case, we
need to expose the compiler error context flag which is now being
controlled by a backend experiment.

This change also does a general cleanup and reorganization of
ExplainSyncOrBuildOutput and its test.

Bug: 352119234
Test: ExplainSyncOrBuildOutputTest
Change-Id: I46b8b7c420d7ad37d958ac76b9e54fb6fa6cb255

GitOrigin-RevId: 027fe21865287bab0d0ce59c47d4bdb7f335fe80