Fix composable function extract test for 2024.2
This functionality works fine in 2024.2, but parts of the platform have
been rewritten to use coroutines. As such it'smuch more complicated to
figure out when the extraction is complete.

I'm rewriting this test so that it's not on the EDT the whole time, and
explicitly waits for a signal that the extraction is complete before
checking the result. This passes both in the mainline and after the
2024.2 merge.

Bug: 352544524
Test: ComposableFunctionExtractableAnalyserTest
Change-Id: I8df8210eb6b6788c67492bf2ae4281905fb2025b

GitOrigin-RevId: 6382596357c23beeb0fae97e8919a6e9b51c7174