[SDK Index] Show notification after sync for blocking issues
... to improve awareness of SDK index issues.

This notification is shown only once per Android Studio session, or when
users fixes all issues and then reintroduce a new one. The notification
shows a link to open the PSD in the Suggestions perspective, where more
details can be seen.

This change is based on commit 85fea338ad49f9322467bf6975e795ca0b987aae
but adds a flag to control wheter or not the notification will be shown.

Bug: 344905371
Bug: 345517435
Test: SdkIndexIssuesSyncListenerTest
Change-Id: I7ef379ded4dc7797728a5ee76af3860e5f561f76

GitOrigin-RevId: 07fb7c4c63e47078077910feb401d2d8b889db05