Fix wear-designer tests after Intellij 2024.2
The main problem originated from a data context lambda not being
recognised. This led to some timeouts as we were no longer able to
get the preview element.

There was also another error in WearTilePreviewDeviceSpecInspectionTest
which was caused by the file some fixes were being applied not being
fully up-to-date when re-running the checks. This CL changes the API
to use fixutre.launchAction instead, which seems more reliable.

Furthermore, this CL changes the setup of
WearTilePreviewDeviceSpecInspectionTest to use
AndroidProjectRule.withAndroidModel() instead of manually installing the
SDK.

Fixes: 352534896
Test: //tools/adt/idea/wear-designer
Change-Id: Ia87da45f49f3ab0125fca34c78975f7efe547290

GitOrigin-RevId: c9a0e82656b70974c6bd8928e261de786d803641