Remove obsolete perf test
This test is running and passing, but isn't actually doing anything. It
was originally meant to validate a certain performance threshold and
publish metrics, but due to changes over time it's no longer finding a
profiler and isn't doing that. On the flip side, it's making merges
harder and is the only usage of this utility in our code base.

Given the test isn't adding anything at the moment, I'm opting to just
remove it.

Bug: 353272351
Test: AndroidLayoutDomTest.testCustomAttrsPerformance
Change-Id: I1c3e96d00a30a3ef3ae8c1c3bcdfbf2a506c9db2

GitOrigin-RevId: ef842e3cab8b2733be94fa4a2f1a989b8ffd1231