Fix testGotoStyleableAttr_frameworkAttr for merge
This test should be validating a framework styleable attribute defined
on a custom view. Since it's a custom view, it should be on the module's
R class, not one from the framework. (And in fact, the next test is the
same thing but for Kotlin, and is doing it correctly.)

This starts failing with the 2024.2 merge because the referenced R class
doesn't exist; pointing to the correct one allows the test to work both
in the mainline and after the merge. (Honestly I have no idea why this
was working to begin with, but at least now the test is validating the
right thing.)

Bug: 353272351
Test: testGotoStyleableAttr_frameworkAttr
Change-Id: I900c213bbb09d9e4d01ba9a2487188e5f3e8131c

GitOrigin-RevId: 518447baf2d9db1d29f704bb67798fa25c21fd5e