Ensure SmartMode is reached by MissingAndroidPluginIssueCheckerTest tests
With IntelliJ 2024.2 the BuildIssueQuickFix.runQuckFix returned Future never completes since this doesn't reach the smartMode given that no sync is triggered on MissingAndroidPluginIssueCheckerTest. For this reason, is required to wait explicitly for the indexes to finish which at the same time dispatches the IDE events queue allowing to reach smartMode and complete the Future.

Bug: 352708863
Test: MissingAndroidPluginIssueCheckerTest
Change-Id: Ic8519c4803ec3378f0777582abd437c6a409aaba

GitOrigin-RevId: 80e6ecf8817c142a0a050a1bf6c6ed87003e4d90