Avoid using project.findProperty() for project isolation
Project.findProperty() tried to find the property by climbing from the
current project to the root project, which violated project isolation.

ProviderFactory.gradleProperty() ignored project level gradle properties
and directly fetch gradle properties on the root level. This is not
considered as violation of project isolation.

As we have moved away from considering the "project level gradle properties"
for quite a long time(since we start to work on cc), the gradle
properties fetched here should be the ones for the build.

Bug: 190016272
Test: ProjectIsolationTest
Change-Id: Ie616292db8fc47b854e40f3427a928baf4ef9b94

GitOrigin-RevId: 82a7159639501dfa365b6463da15f8b15e5720f9