Address flaky manifest test
This test has a very low flake rate, and seems to hit across many of its
various test methods. My theory is that when the manifest is updated,
there's an uncommon race condition where we wait for manifest updates
*before* those updates are queued, causing us to continue test execution
prematurely.

Dispatching EDT events before waiting should address that, since the
queueing happens in response to events that are coming on the EDT and
will already be waiting to go by the time we ask them to be dispatched.

Test: AndroidManifestIndexQueryUtilsTest
Fixes: 352014593
Change-Id: I57f9461bd21e6a4062cf2683d371d9b43ab74ec2

GitOrigin-RevId: 5ce323fbac0cdd169449603bd6fe7b0ac0ffdf37