Adapt tryEvaluateConstant to K2 AA changes.
Analysis API in 2.0 removed the CONSTANT_LIKE_EXPRESSION_EVALUATION mode
for constant evaluation. This mode was used by AndroidKtPsiUtils to
evaluate the constant value of variable-reference expressions.

As a workaround, manually resolve property reference expressions to the
initializer of the property before performing constant evaluation. This
covers most, but not all, cases that CONSTANT_LIKE_EXPRESSION_EVALUATION
would handle. (The remaining cases will be treated as non-constant,
which callers already need to be prepared to handle.)

Bug: 352544083
Test: //tools/adt/idea/android-common:intellij.android.common_k2_tests
Change-Id: Ifaaf83ba77f76a325984645c5dda0f7ba5e74b0d

GitOrigin-RevId: 8fbde40d6359c35f8c23660b83ed1667a8c69247