Remove synchronous rendering from LayoutlibSceneManager
When an animation is being played in motion editor, its design
surface is in a "synchronouslyRendering" mode, which is used for
turning off some parts of it, like the errors panel, to make
re-renderings happen faster. However, it's not necessary to avoid
triggering some listeners in the background in LayoutlibSceneManager.

This change removes this redundant code.

Bug: 335424569
Test: manually tested that nothing changed in motion editor
Change-Id: I8d68f63ab6394fd1b55e2d0e0a0778e76defcdff

GitOrigin-RevId: 22e2ccaec9b54376dc236d76a5338e592ba9a266