Clean up AndroidVirtualDevice
This file seems to have been auto-converted to Kotlin without much
cleanup.

Instead of passing Map<String?, RemotePackage>?, just pass a
Collection<RemotePackage>. We never use the map keys (and they are
present in the RemotePackage already).

Make a primary constructor that just takes the AndroidVersion to make it
clearer that the only thing we do with the remotePackages input is
extract an AndroidVersion. That also lets us avoid creating a bunch of
fake packages in tests just to wrap an AndroidVersion.

Bug: n/a
Test: existing
Change-Id: Icccce300927a8f00d0642c536715d587e5463bb7

GitOrigin-RevId: fd0b5bd6da25df9101b4dbfacdd33a7424c8a870