Do not prioritise Android configs over Gradle's
When a test can run both as a Gradle test and as a instrumented
test, then we should show give the choice to run both of the
options rather than preferring one over the other. This is
applicable for shared test sources (ex: module scope), or
multiple available tests environments.

We also need to ensure that by doing so, and not making the android
RC preferred (replace) to the Gradle RC, to change the test
runner (org.jetbrains.plugins.gradle.settings.TestRunner) to be
based on every test type rather than GRADLE in order to not end up
with GRADLE configurations only suggested.

Bug: 174357889
Test: added
Change-Id: I66f83b04270292537fe74c181c3851948f7eaa61

GitOrigin-RevId: 82cd5ec45cd62e3f7f67d9ac75f16659be293e8f