Introduce WearHealthServicesPanelManager to handle
... the creation and storage of WHSPanels.

EmulatorViews are disposed as soon as a user switches the device view
in the running devices tool window.

We want to be able to keep track of the same panel even if the emulator
view is disposed of.

This CL uses the EmulatorController as the parent disposable. This will
allow us to better handle showing error messages either on the panel
itself or to display it in a notification.

Bug: 346275910
Test: WearHealthServicesPanelManagerTest
Change-Id: I07961041a6178b362dd5cadb21612311121bd0e6

GitOrigin-RevId: 2b8952590898855741a081b376250243b88a7bff