Remove unnecessary update queue
As far as I can tell, the update queue serves to minimize the number of
`Runnable`s that are placed on the update executor. It does that by
having any `Runnable` that's executing run every single pending action
before completing, and any new actions that are added only queue a new
`Runnable` when nothing is currently executing.

This appears to be an optimization, since the executor is already
running only a single thread. Queuing a new `Runnable` for each action
should operate in the same manner.

More importantly, this simplifies the logic and makes the transition to
coroutines easier.

Bug: 354682619
Test: ResourceFolderRepositoryTest
Test: ResourceFolderRepositoryBackgroundActionsTest
Change-Id: I5907fec04c81ae9b32a2662518a1b5c8c69a7ef6

GitOrigin-RevId: d8ea3114fadafacd450b901938bb2d770d344524