Flag to use flat list of dependencies on the IDE
We intended to use the graph in the IDE, but it comes with a large
performance cost and is currently unused, so switching to a flat list
of dependencies instead for each variant.

This is expected to improve the performance of Android model building
by not asking for the resolution result (i.e. the graph), therefore
avoiding additional work.

However, there is currently a regression with Gradle 8.7 that introduces
some lock contention, which means the improvements will not be as
visible until we land a version with a fix (I've confirmed this is the
case by "fixing" the bug locally on Gradle)

Test: Covered by snapshot tests
Test: UnresolvedDependenciesReporterIntegrationTest
Bug: 354906596
Change-Id: I590368de8d9917b9ec4da71011f15a9ae56d8841

GitOrigin-RevId: 5fd81c7e04d16a28d34e9928a9aa6a2d34654d27