Use coroutines instead of pooled threads
This change is meant to be functionally identical to the previous logic.
All the actions that were previously run on an executor are now being
done in a coroutine context, which avoids allocating a lot of extra
threads. The `Channel` queues are used to ensure only a single action is
run at a time for each queue, and that the actions are run in the order
they're submitted.

Switching back to an indicator process via `coroutineToIndicator` is
technically unsupported (it's an @Internal API meant for helping with
platform migration). Technically this is a similar migration - it is
allowing us to migrate the submission of actions to coroutines without
currently modifying the internal scanning logic. That can still be done
at any point in the future, but is not necessary for this bug to fix the
thread allocation issue.

Fixes: 354682619
Test: ResourceFolderRepositoryTest
Test: ResourceFolderRepositoryBackgroundActionsTest
Change-Id: I21e9ffca733d3d10b3768fda0b85d69685cb0760

GitOrigin-RevId: 78821a99eca24bb41b062ec7d61d99f3c17b85ba