Stop handling dumb mode in NavInfoFetcher
NavInfoFetcher currently checks for dumb mode and returns null, in an
effort to avoid accessing an index when it's not available. However, in
certain cases the platform will opportunistically call this logic while
in dumb mode, and it takes the null result (as opposed to a thrown
exception) as an indication that all is well, and there's no actual data
to be returned. This leads to nav classes failing to resolve in the
editor.

Similar checks in data and view binding light classes have already been
removed with success. Even though the code can now throw, platform code
higher up the stack should generally be resilient to that.

Unfortunately, this case is materially different than the data and view
binding classes, in that the nav info is used to generate synthetic
Kotlin (K1) descriptors. Those are unique in our codebase to this
scenario. So it's possible there are parts of the platform that do *not*
handle these index accesses correctly, and an exception may bubble up.
For that reason, I've put this code behind a flag so that it can be
toggled off quickly if need be.

A note: a recent change included a log line that helped identify some of
the stacks where this IndexNotReadyException may occur. I have reviewed
manually many of those stacks, and believe the ones I've looked at
should be safe/resilient to the exception.

Bug: 352610609
Test: NavInfoFetcherTest
Change-Id: I3298b8363433334d5a91badedf32252353a8ee59

GitOrigin-RevId: 4bc15fbc1adbb35fe69a1b176cfce802aa4aaeaa