Fix button not being disabled when the emulator goes offline
When a DeviceHandle becomes ready we update the main toolbar to ensure
the button becomes enabled. However, whenever the device was
disconnected, we would not update the toolbar. This would lead to the
button sometimes staying enabled even though it was not supposed to be.

This CL ensures that we update the toolbar when we detect that the
device goes offline.

Fixes: 346506392
Test: Updated EmulatorAdbReadyServiceTest
Change-Id: I5ed36bc1e5fab712b1ec98edd57c31652fec1023

GitOrigin-RevId: 142f2742d816a051438cc3e8fe991725c5b3b412