Fix flaky test and re-enable ignored test
The flakiness was introduced by I477341a9ba2464442ee3c which prevented
override values to be set outside an ongoing exercise.

This CL ensures the exercise is started before overriding the values.

The test that was previously ignored might need to be revisited as it
exposes a problem:

We allow overriding a value for a disabled capability and show that
disabled capability's value. It sort of makes sense to show the value
as it's what is returned by the synthetic HAL API and because the
user set that previously. However, it can be a bit misleading as the
sensor is supposed to be disabled. It might be best to not show the
value nor the input box.

However, this is out of scope of this CL, as this CL aims at just
removing flakiness.

Bug: 342411390
Fixes: 355505515
Test: Updated WearHealthServicesPanelTest
Change-Id: I9507c53a3d032d1d1d7b978c63cd32882a88758c

GitOrigin-RevId: 28404dead1bd576b3a56ebf16693393e7c4424b3