Fix WearTilePreviewDeviceSpecInspectionTest for k2
The issue was coming from WearTileDeviceSpecInjectionContributor when
applying the fix. WearTileDeviceSpecInjectionContributor was failing
due to K2, for the same reason as ComposeDeviceSpecInjectionContributor
in Ic0a2005a291b2cf0b25c82a: we need to allow running an analysis in
the EDT and write action as there is no other choice for the
injection contributor.

This CL also introduces a test for
WearTileDeviceSpecInjectionContributor to make it easier to catch the
issue.

Fixes: 355606883
Test: WearTilePreviewDeviceSpecInspectionTest and
 WearTileDeviceSpecInjectionContributorTest with k2 enabled
Change-Id: Id06929fef9bc749ee0b71a48718b72675bef9f62

GitOrigin-RevId: 9030e26f186ef7f1831c83dfa0b1a1813edd0e54