Add oem-labs module
Importantly, don't add it as a dependency of the root module, since we
don't want it to be included in Studio by default.
Note also that this module is entirely temporary: it's only needed
to enable oem lab support while the feature is in development on the
server side: once it's done it will be available via the normal
direct access APIs and anything else needed will be added directly in
the normal direct access modules.

Test: none
Bug: none
Change-Id: I8ec237e9b5777113b33b59a01ab2d0d51bb67674

GitOrigin-RevId: a9ed3bece8da46f25b055f0132dc8dc9e5bba63d