Extract intellij.android.common tests to a separate module
This is being done to support ag/28488225. In that CL we are adding
a new module called mendel-experiments, and exposing a new extension point there. The extension point needs to be tested in
intellij.android.common tests. From my observations, it seems like
this only works if the test is in a test-only module. The test
module also needs to include all of the necessary dependencies
for the Android plugin to load, so that go/plugin-xml-2023 kicks in
and makes the extension point visible in the test. That's why
ag/28488225 adds intellij.android.testFramework as a dependency
in this new test module.

Bug: 342183628
Test: existing
Change-Id: I8dd141ec7a53824cb398662bc04dacf188fc890e

GitOrigin-RevId: 00a9dddf0e840d12ae7da62ae4da0d0621255e56