[AUA] Add refactoring for useAndroidX default change
Starting with AGP 9.0 the default value of android.useAndroidX will
change from false to true. The behavior of projects updated from older
versions will change if the property is not explicitly set to false.

This change adds UseAndroidXDefaultRefactoringProcessor to explicitly
set the property when updating projects from pre 9.0 that do not set it.
This change also adds AbstractBooleanPropertyDefaultRefactoringProcessor
from which new processors can be created for changes in the default
values of gradle properties.

Bug: 353973156
Test: UseAndroidXDefaultRefactoringProcessorTest, ComponentTrackerTest
Change-Id: I4f1dc17044f3b527605a99a2e76291569c0788f7

GitOrigin-RevId: 9b5bb08df456c3da273fc54e2fddc6301fb8786f