Don't run ScriptDefinitionsManager.getAllDefinitions in tests
This settings adjustment is not necessary for tests.
At the same time running this results in flakiness in tests.
Different services getting initialised, including GradleScriptListener.
With 242 platform merge this process starts happening asynchronously
which sometimes results in init happening during project disposal as the
test has already finished.
In some of such cases it can happen that the listener is installed but
never removed, because of the bug in VirtualFileManagerImpl. Which
results in test flakiness when some other test triggers this broken
listener holding to disposed project of previous test.

Basically any test creating gradle files can be affected by this
flakiness.

Bug: 356225801
Test: fixing tests flakiness
Change-Id: I6ed7674c1db7badf66b8f067d07433364bce57d0

GitOrigin-RevId: 028e1109f39a611d2f44405f01e1049c936d4671