Fix layout switch UI
This action was not reflecting the last state because we were reading
the state from a different place. Because there is a lag between setting
the value and being updated in the surface, the action would reflect the
wrong value.
This change updates the action so the information is read from the same
source, ensuring that it is always up to date.

Bug: 357035207
Test: Manually verified
Change-Id: Ib82215590c252284a6b5f41fe098c858cb56436a

GitOrigin-RevId: ddedb2d69b4f1e19b482f2a4b7d8273fce670d56