Enable skipIfRuntimeNotFound option for compose compiler
Loading the compose version class from the compose runtime will
cause an exception for a non-compose module because of the missing
dependency on compose runtime. This commit enables
`skipIfRuntimeNotFound` option to prevent the compose compiler
plugin from throwing an exception when the compose version class is
not found, which is for non-compose modules.

Bug: 339317719
Test: Manually tested a compose app with a non-compose module
Change-Id: I8c906bc60ef0810d82a436e543f855b26cea83bd

GitOrigin-RevId: 9073241d4b28ed298eb5e35ac234aa16495f6847