[AUA] Add refactoring for library constraints
Users have been able to use an experimental flag to turn off library
alignment, this flag was renamed in a previous AGP version and starting
with AGP 9.0 the default value changed to true. We decided to keep the
refactoring of this property simple and only treat it as a change in
default, ignoring the experimental property.

This change adds a refactoring processor to add the old default when
updating from a version older than 9.0 to 9.0+ if the property is not
already used.

Bug: 356892144
Test: ComponentTrackerTest, ExcludeLibraryDefaultRefactoringProcessorTest
Change-Id: I0a300fa26ed7d27127026a30ad24a18d4e602f45

GitOrigin-RevId: 7c3865df1c68dad3d1201178ea20f6530619ae91