Fix PlatformIntegrationTest running over 5 minutes
testCorrectSyncEventsPublished_dataImporterCrashesAfterSuccessfulOpen
was constantly running for more than 5 minutes.
The reason was the combination of several factors:
1) TestGradleSyncListener.await silently ignored await timeout and
happily proceed after. Change to explicitly failing any test on this.
2) TestGradleSyncListener was not invoked on importer crash because
IdeaSyncPopulateProjectTask was awaiting Exception in catch, but in this
case exception was wrapped in TestLoggerAssertionError and did not
trigger this codepath. Fix by catching all throwables, it seems more
correct here and we rethrow it anyway.

Bug: 360006227
Test: improving tests.
Change-Id: I3fbead0517446dec35efc5b4e743a4b442496f67

GitOrigin-RevId: c611e34d1f863631be0e72c65ffd79fc3492349d