Show error if standalone profiler has task-ux on
In Standalone Profiler, we disable task-based UX by setting the
JVM property at the command line:

  -Dprofiler.task.based.ux=false

inside the standalone profiler launch script. However, if the user
goes to the 'Studio Flags' and explicitly sets the flag to 'true'
(rather than clicking on the 'reset to default', which is also
'true'), then that explicit user selection takes precedence over the
command line JVM property.

This change detects this case and warns the user to not do that.

Test: n/a
Bug: 338285051
Change-Id: I0535c1134d8c6d8b55e05e543b39bda59098cd77

GitOrigin-RevId: 1b05e340cb1bb5ab2b13a40c25e0aa977666b9b8