Remove a usage of isHolderModule from core code
Make the project system (through a Token) responsible for answering
whether a particular module should be accepted in a run configuration
editor.

(I suspect that this involves the same underlying concept as
Project.getAndroidModulesForDisplay(), but we don't want to generate a
list of modules each time this is called just to be able to filter out
those which are not present.)

While we're here, generate the resolve scope for the run configuration
editor using the same mechanism.

Bug: 147586218
Test: existing
Change-Id: I3fb0a9cf6be7bd1cdfaf1e81cd4cf3364d593aca

GitOrigin-RevId: 59fc96c4a142a51121c0fc1857617130a917ad44