Move the Proguard/R8 ResolveScopeEnlarger
This enlarger smooths over the fact that for
Gradle (module-per-source-set) projects, the proguard file is in the
holder module, while the entities it refers to are typically in the
main source-set module.  This is inherently applicable only to Gradle
projects, so move it to the Gradle project system, and add a check
that the project system is in fact a GradleProjectSystem before
enlarging the scope.  (For non-Gradle projects, the scope added to the
inherent scope is tautologically the same module, but we might as well
not do that).

Bug: 147586218
Test: existing
Change-Id: Icc83ac314f7d42f0f4c85737a7236aea16f504d9

GitOrigin-RevId: b81f8e4dc55218cdc308a5131573980305a1c5e2