Change logic of opening/closing animation inspector
The idea behind all changes that we should have one instance of
ComposeAnimationPreview for all Application, so if user tries to open a
new animation preview we are closing existing one.

For that we send a messageBus with newInspectorIsOpening, each not
disposed restores previous mode (equivalent for pressing stop button).

Before that we used ComposeAnimationInspectorManager.kt, apparently with
time logic inside it degraded and the class contained many errors, like
closing inspector not belonging to file on file dispose, not closing
extra inspector at all and so on.

Extract tests for ComposeAnimationSubscriber into separate file, leaving
in ComposeAnimationPreviewTest only tests directly working with
ComposeAnimationPreview.

Fixes: 355391435
Test: ComposePreviewTest.testAnimationInspector, ComposeAnimationPreview
Change-Id: I7bee27aa219eac5cb1f5cc4db6fdbf2059794ad3

GitOrigin-RevId: 73832b37b80f287ef6e2a4d3c9c135ebcc96275c