Report build events on already running builds
Make `BuildSystemFilePreviewServices` report build started/finished
events on builds that were already running when a listener was
subscribed. However, report the scop of a completed build as empty
indicating that no files are guaranteed to be built since the build
started event.

This is supposed to be a temporary state until we can transition to
runtime/rendering artifact availability listeners.

Bug: 353981758
Test: GradleBuildSystemFilePreviewServicesTest
Change-Id: Ifc243d6c486029fbef1efb35fc171dcad433a238

GitOrigin-RevId: 09df304dcf553b37b6c94cefe8b3576a2aa65a75