Translate from holder to sourceSet modules using the project system
The right thing to do for non-Gradle projects is not obviously the
same as saying that such a module should not have instrumented tests.

Preserve the existing behaviour (both for Gradle projects and for
non-Gradle projects), while making it more obvious how and where the
existing behaviour could / should be overridden if needed.

Bug: 147586218
Test: existing
Change-Id: Ib6d6396c9321d5c4a70c3d27fadaa133c204390e

GitOrigin-RevId: f7b2652fa2614cf6e9324a69b051c2dddc43a97d