Remove "Custom" capability preset and replace combobox
... with a "Load preset" button with two preset options: "all" and
"standard".

This CL also changes the behaviour to not "track" the current preset
but to allow the user to "load" a preset instead. This should help
simplify the concept for the user and cause less confusion around
what impact the currently selected preset had.

This CL also changes when the WearHealthServicesStateManagerImpl is no
longer considered "Initializing". Before this CL, the status would be
set to "idle" after the preset combobox was initialized and the preset
flow was set to the selected item. With this CL, we now instead consider
the status to be "idle" once a serial number has been set in the
state manager.

Finally, this CL adds a tooltip text on the "Load preset" button to
add some additional information for the user on the purpose of the
button aiming at reducing user confusion.

Fixes: 346685480
Test: updated tests
Change-Id: Ic9251779a5c74da40c3d589c3533e61ef254dbbf

GitOrigin-RevId: da7377e052757fd2a23aedc17314782b50cd2d98