Remove circular dependency between SceneViewPanel and NlDesignSurface
The issue seems to be caused by the following events:

We initialize SceneViewPanel in the NlDesignSurface constructor.
SceneViewPanel takes a NlDesignSurfacePositionableContentLayoutManager
in the constructor, and this class calls NlDesignSurface.onLayoutUpdated
a collect inside its own constructor.

So this Cl moves collector on surface after surface was initialized
fully

+ remove not necessary parentDisposable param in DesignSurface, we
should create scope based on `this` and NOT on parent

Fixes: 328203421
Test: NlDesignSurfacePositionableContentLayoutManagerTest
 + existing tests don't throw exception
Change-Id: Icc7a8fd39119069ce4d0275c1509c321ac9d9619

GitOrigin-RevId: 341fc620464ca76284ddc7ca580e7b59298654b6