Use the window bounds in standalone mode
The layout bounds received from the view and compose inspector are
in screen coordinates which is exactly what is needed for RDW.

In standalone mode we need to adjust the coordinates if the window
is not in the top left corner of the screen. Example: an app in
the lower half when running in split mode on the device.

Use the origin of the window bounds rectangle instead of the screen
coordinates for all view and compose nodes and for Quad coordinates
from the view agent.

The compose agent sends Quad coordinates relative to the window.
We could change the compose agent to send screen coordinates but
we would have to support the older window coordinates as well.
For now: keep the window Quad coordinates in the compose agent.

In order to get scrolling to work correctly we also need to change
the apparent screen size to the window size.

Bug: 361367506
Test: AppInspectionTreeLoaderTest
Change-Id: I2a9731f4784e698e891eb82e92c37c0595f3c99b

GitOrigin-RevId: 3ff7fbe3184afba027be3ce731de084ddd8926b5