Extract a method to click on the apply button
There is a flow that needs to be consumed whenever a user clicks on the
apply button. The new method aims at reducing errors due to not
consuming that flow.

This CL also changes the JComboBox.hasLabel method to test for label
equality instead of just using "contains". This change highlighted some
tests that were not consuming the "apply" flow.

Bug: None
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: I23043dd237525ecbd6de899ef75be4144a9a1e8d

GitOrigin-RevId: 364b9ceb8252e58da61fd6f8223cc9142791ca12