Execute Restore After Deploy
This allows restoring on a device when the app is not yet installed.

Moving the "restore" step after the "deploy" step exposed that backup
is not supported for Instant Apps. Attempting top backup an instant app
results in:
```
 $ adb shell bmgr backupnow com.example.instantapp
Running incremental backup for 1 requested packages.
Package com.example.instantapp with result: Backup is not allowed
Backup finished with result: Success
```

So this cl also disables the `RestoreRunConfigSection`

Bug: 348406593
Test: Manually
Change-Id: I53e91a707dfc43e5b67003d1e7cc39ee72198e8e

GitOrigin-RevId: d9e78be24bd1e2e5929cb2da348902228c594fe4