fixup! [android] (Ladybug) Merge AS 2024.2.1 Canary 7: Attempt to fix hang in com.android.tools.idea.navigator.AndroidProjectViewTest tests
`testGeneratedResources` and `testGeneratedAssets` are hanging indefinitely because Gradle sync never completes.

Sync never completes because in test setup the project is not linked to Gradle.
When sync is invoked, the following sequence is executed:
1. com.android.tools.idea.testing.AndroidGradleTestCase.requestSyncAndWait()
2. com.android.tools.idea.testing.AndroidGradleTests.syncProject
3. com.android.tools.idea.gradle.project.sync.GradleSyncInvokerImpl.requestProjectSync
4. com.android.tools.idea.gradle.project.sync.idea.GradleSyncExecutor.sync

In the last step, sync fails to setup project because no android project candidates were found.

In Android Studio Gradle sync for test works because there is a fallback to attempt to link the Gradle project.

GitOrigin-RevId: 3cb7fc1c7a43f4ac58be090f43354ed27defeb27